backend/requirements.txt.txt
fastapi
uvicorn
sqlalchemy
sqlite3
databases
passlib[bcrypt]
pyjwt
python-dotenv
